fix: correct rumble_def and values field handling

+++ packages/fable-data/src/def/defs/rumble_def.rs
@@ -5,7 +5,9 @@ def_struct! {
     /// `CRumbleDef` — C++ `CRumbleDef`.
     #[derive(Debug, Clone, PartialEq)]
     pub struct RumbleDef {
-        "QuakeIntensities" => pub quake_intensities: VecMap<i32, i32>,
-        "QuakeDurations" => pub quake_durations: VecMap<i32, i32>,
+        // C++: `std::map<EQuakeStrength, float>` / `std::map<EQuakeLength, float>`
+        // (tcd_rumble.hpp) — the values are floats, not ints.
+        "QuakeIntensities" => pub quake_intensities: VecMap<i32, f32>,
+        "QuakeDurations" => pub quake_durations: VecMap<i32, f32>,
     }
 }
